requirements.txt
gdown==5.2.0
pillow==11.1.0
streamlit==1.41.1
openai==1.76.0
transformers==4.51.3
onnxruntime==1.22.0
faiss-cpu==1.11.0
pdfminer.six